+++ src/main/java/org/jabref/preferences/GuiPreferences.java
@@ -3,18 +3,27 @@ package org.jabref.preferences;
 import java.nio.file.Path;
 import java.util.List;
 
+import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleObjectProperty;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+
 public class GuiPreferences {
-    private final double positionX;
-    private final double positionY;
-    private final double sizeX;
-    private final double sizeY;
+    private final DoubleProperty positionX;
+    private final DoubleProperty positionY;
+    private final DoubleProperty sizeX;
+    private final DoubleProperty sizeY;
 
-    private final boolean windowMaximised;
+    private final BooleanProperty windowMaximised;
 
-    private final boolean shouldOpenLastEdited;
-    private List<String> lastFilesOpened;
-    private Path lastFocusedFile;
-    private double sidePaneWidth;
+    private final BooleanProperty shouldOpenLastEdited;
+    private final ObservableList<String> lastFilesOpened;
+    private final ObjectProperty<Path> lastFocusedFile;
+    private final DoubleProperty sidePaneWidth;
 
     public GuiPreferences(double positionX,
                           double positionY,
@@ -24,65 +33,118 @@ public class GuiPreferences {
                           boolean shouldOpenLastEdited,
                           List<String> lastFilesOpened,
                           Path lastFocusedFile, double sidePaneWidth) {
-        this.positionX = positionX;
-        this.positionY = positionY;
-        this.sizeX = sizeX;
-        this.sizeY = sizeY;
-        this.windowMaximised = windowMaximised;
-        this.shouldOpenLastEdited = shouldOpenLastEdited;
-        this.lastFilesOpened = lastFilesOpened;
-        this.lastFocusedFile = lastFocusedFile;
-        this.sidePaneWidth = sidePaneWidth;
+        this.positionX = new SimpleDoubleProperty(positionX);
+        this.positionY = new SimpleDoubleProperty(positionY);
+        this.sizeX = new SimpleDoubleProperty(sizeX);
+        this.sizeY = new SimpleDoubleProperty(sizeY);
+        this.windowMaximised = new SimpleBooleanProperty(windowMaximised);
+        this.shouldOpenLastEdited = new SimpleBooleanProperty(shouldOpenLastEdited);
+        this.lastFilesOpened = FXCollections.observableArrayList(lastFilesOpened);
+        this.lastFocusedFile = new SimpleObjectProperty<>(lastFocusedFile);
+        this.sidePaneWidth = new SimpleDoubleProperty(sidePaneWidth);
     }
 
     public double getPositionX() {
+        return positionX.get();
+    }
+
+    public DoubleProperty positionXProperty() {
         return positionX;
     }
 
+    public void setPositionX(double positionX) {
+        this.positionX.set(positionX);
+    }
+
     public double getPositionY() {
+        return positionY.get();
+    }
+
+    public DoubleProperty positionYProperty() {
         return positionY;
     }
 
+    public void setPositionY(double positionY) {
+        this.positionY.set(positionY);
+    }
+
     public double getSizeX() {
+        return sizeX.get();
+    }
+
+    public DoubleProperty sizeXProperty() {
         return sizeX;
     }
 
+    public void setSizeX(double sizeX) {
+        this.sizeX.set(sizeX);
+    }
+
     public double getSizeY() {
+        return sizeY.get();
+    }
+
+    public DoubleProperty sizeYProperty() {
         return sizeY;
     }
 
+    public void setSizeY(double sizeY) {
+        this.sizeY.set(sizeY);
+    }
+
     public boolean isWindowMaximised() {
+        return windowMaximised.get();
+    }
+
+    public BooleanProperty windowMaximisedProperty() {
         return windowMaximised;
     }
 
+    public void setWindowMaximised(boolean windowMaximised) {
+        this.windowMaximised.set(windowMaximised);
+    }
+
     public boolean shouldOpenLastEdited() {
+        return shouldOpenLastEdited.get();
+    }
+
+    public BooleanProperty openLastEditedProperty() {
         return shouldOpenLastEdited;
     }
 
-    public List<String> getLastFilesOpened() {
+    public void setOpenLastEdited(boolean shouldOpenLastEdited) {
+        this.shouldOpenLastEdited.set(shouldOpenLastEdited);
+    }
+
+    public ObservableList<String> getLastFilesOpened() {
         return lastFilesOpened;
     }
 
-    public GuiPreferences withLastFilesOpened(List<String> lastFilesOpened) {
-        this.lastFilesOpened = lastFilesOpened;
-        return this;
+    public void setLastFilesOpened(List<String> files) {
+        lastFilesOpened.setAll(files);
     }
 
     public Path getLastFocusedFile() {
+        return lastFocusedFile.get();
+    }
+
+    public ObjectProperty<Path> lastFocusedFileProperty() {
         return lastFocusedFile;
     }
 
-    public GuiPreferences withLastFocusedFile(Path lastFocusedFile) {
-        this.lastFocusedFile = lastFocusedFile;
-        return this;
+    public void setLastFocusedFile(Path lastFocusedFile) {
+        this.lastFocusedFile.set(lastFocusedFile);
     }
 
     public double getSidePaneWidth() {
+        return sidePaneWidth.get();
+    }
+
+    public DoubleProperty sidePaneWidthProperty() {
         return sidePaneWidth;
     }
 
-    public GuiPreferences withSidePaneWidth(double sidePaneWidth) {
-        this.sidePaneWidth = sidePaneWidth;
-        return this;
+    public void setSidePaneWidth(double sidePaneWidth) {
+        this.sidePaneWidth.set(sidePaneWidth);
     }
 }