+++ src/main/java/org/jabref/preferences/PreferencesService.java
@@ -181,11 +181,7 @@ public interface PreferencesService {
 
     RemotePreferences getRemotePreferences();
 
-    void storeRemotePreferences(RemotePreferences remotePreferences);
-
     ProxyPreferences getProxyPreferences();
-
-    void storeProxyPreferences(ProxyPreferences proxyPreferences);
 
     //*************************************************************************************************************
     // CitationKeyPatternPreferences
@@ -302,8 +298,6 @@ public interface PreferencesService {
     //*************************************************************************************************************
 
     GuiPreferences getGuiPreferences();
-
-    void storeGuiPreferences(GuiPreferences guiPreferences);
 
     void clearEditedFiles();